package/dropbear: Add separate configuration option for dropbear server
Currently, the dropbear package installs both client and server components.
This means that when we only want the client binaries, we also get the server,
which is run from init.

Even though it's a multi-call binary (the client and server exist in the same
executable), we can define which parts are compiled-in. We'd also like to
selectively install the links and init scripts.

This change introduces a separate configuration for the dropbear server.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
[updated for 2016.05]
Signed-off-by: Joel Stanley <joel@jms.id.au>