linux-headers: allow specifying 4.9 as kernel header version
Even though 4.9 is not released yet, some people may want to build a
system using the 4.9-rc kernel, and point to the kernel sources as the
kernel headers to use for the toolchain.

This also allows users to select a 4.9 custom kernel and have it build
GCC with headers from that kernel.

This is the 4.9 equivalent of Fabio's d3ac2bc61680 (linux-headers: allow
specifying 4.8 as kernel header version).

Signed-off-by: Joel Stanley <joel@jms.id.au>