package/linux-headers: add support for Linux 5.0 kernel headers
This sets 5.0 to be the default, moving from 4.19 which was used instead
of 4.20 for the LTS branch.

We need to update linux/linux.hash as part of this commit, because
package/linux-headers/linux-headers.hash is a symlink to
linux/linux.hash.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 6c6137ab2b0be6278b43ae187e84b1d56bcd2b97)
Signed-off-by: Joel Stanley <joel@jms.id.au>